+++ com.unity.renderstreaming/Runtime/Scripts/StreamReceiverBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 12caa12a46f768a4ebde5979f000a8c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
