+++ com.unity.renderstreaming/Runtime/Scripts/StreamSourceBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: efb037253f83cd74d8da87f533785bf6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
